+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+== 0.5.6 Added loading animation
+
+== 0.5.5 Fixes for stability and speed
+ * Make sure that initilization is run only once 
+ * Make sure that DOM watchers are initialized only once, and after toolbar installation
+
+== 0.5.4 New purple icon
+
 == 0.5.3 Auto-refresh on on-board updates
  * Refresh after moving a card to another column (and other actions made directly on the board page), issue #7